requirements.txt
clickhouse-driver>=0.2.0
clickhouse-connect>=0.7.0
pandas>=1.5
pyarrow>=18.1.0
apache-airflow>=2.3
python-dotenv>=0.21.0
matplotlib>=3.8
seaborn>=0.12
plotly>=5.24
fastapi>=0.115
uvicorn>=0.30
streamlit>=1.37
strawberry-graphql[fastapi]>=0.262
openai>=1.30
langgraph>=0.2
langchain-openai>=0.2
langchain-core>=0.3
arize-phoenix>=4.0
opentelemetry-sdk>=1.25
opentelemetry-exporter-otlp-proto-http>=1.25
openinference-instrumentation-langchain>=0.1
mlflow>=2.12